requirements.txt
numpy~=1.19.5
torch~=1.8.1
tensor2tensor~=1.15.7
tokenizers~=0.10.3
transformers==4.6.1
wandb~=0.11.1
tqdm~=4.61.0
pandas~=1.1.5
sentencepiece==0.1.95
pytorch-nlp==0.5.0
